created helpers to manage files

+++ .reaxpress/commands.js
@@ -14,34 +14,41 @@ const mkdirSync = (path) => {
   }
 };
 
+const helper = require('./helper');
+
 module.exports = {
 
-  create: (arg) => {
-    console.log(arg);
+  create: (name) => {
+    // bail if route exists
+    if (helper.checkIfRouteExists(name)) {
+      return false;
+    }
+
+    console.log(name);
     // update app.js
     fs.readFile(appFile, 'utf8', (err, data) => {
       if (err) {
         return console.log(err);
       }
-      let modified = data.replace(/#route-def/g, `#route-def\nconst ${arg} = require('../routes/${arg}');`);
-      modified = modified.replace(/#route-mount/g, `#route-mount\nrouter.use('/${arg}', ${arg});`);
+      let modified = data.replace(/#route-def/g, `#route-def\nconst ${name} = require('../routes/${name}');`);
+      modified = modified.replace(/#route-mount/g, `#route-mount\nrouter.use('/${name}', ${name});`);
       fs.writeFile(appFile, modified, 'utf8', (err) => {
         if (err) return console.log(err);
         console.log('successfully added route to app file.');
       });
     });
     // add react components if they do not already exist
-    const reactDirPath = path.join(__dirname, '..', 'src/react/', arg);
+    const reactDirPath = path.join(__dirname, '..', 'src/react/', name);
     const reactFile = path.join(reactDirPath, 'index.jsx');
     mkdirSync(reactDirPath);
-    const componentCode = template.component(arg);
+    const componentCode = template.component(name);
     fs.writeFile(reactFile, componentCode, 'utf8', (err) => {
       if (err) return console.log(err);
       console.log('successfully created react component file.');
     });
     // create route file
-    const routeFile = path.join(__dirname, '..', 'routes', `${arg}.jsx`);
-    const routeCode = template.route(arg);
+    const routeFile = path.join(__dirname, '..', 'routes', `${name}.jsx`);
+    const routeCode = template.route(name);
     fs.writeFile(routeFile, routeCode, 'utf8', (err) => {
       if (err) return console.log(err);
       console.log('successfully created route file.');
@@ -51,7 +58,7 @@ module.exports = {
       if (err) {
         return console.log(err);
       }
-      const modified = data.replace(/entry: {/g, `entry: {\n    ${arg}: './src/react/${arg}',`);
+      const modified = data.replace(/entry: {/g, `entry: {\n    ${name}: './src/react/${name}',`);
       fs.writeFile(webpackConfigFile, modified, 'utf8', (err) => {
         if (err) return console.log(err);
         console.log('successfully added entry to webpack config file.');
